Menu Gestao mobilidade

+++ MobilidadeEletrica.h
@@ -7,12 +7,12 @@ struct MobilidadeEletrica {
     char tipo[50];
     char estado[50];
     char localizacao[50];
-    int bateria;
+    double bateria;
     float preco;
      MobilidadeEletrica *proximo;
 } ;
 
-MobilidadeEletrica* CriarMobilidadeEletrica(int id, char* tipo, char* estado, char* localizacao, int bateria, float preco);
+MobilidadeEletrica* CriarMobilidadeEletrica(int id, char* tipo, char* estado, char* localizacao, double bateria, float preco);
 MobilidadeEletrica* CriarMobilidadeEletricaFixa(void);
 
 #endif /* MOBILIDADEELETRICA_H */
